Create user controllers and views
The controller allows users to create new posts and
also display.
An authenticated user is the only one allowed
to create a new post. Upon authentication, the user is
able to see the post title as well as the author of the title

A user who is not authenticated can only view the post title and not
the author of the post.